+++ crm-data-components/hsproject.json
@@ -1,5 +1,5 @@
 {
-  "name": "example-crm-components",
+  "name": "crm-data-components",
   "srcDir": "src",
   "platformVersion": "2023.2"
 }
