+++ crm-data-components/src/app/app.json
@@ -1,7 +1,7 @@
 {
-  "name": "CRM Components React App",
+  "name": "CRM Data Components App",
   "description": "This is an example of private app that shows 2 custom cards. It demonstrates usages of CRM Components",
-  "uid": "with_crm_components_app",
+  "uid": "crm_data_components_app",
   "scopes": [
     "crm.objects.contacts.read",
     "crm.objects.contacts.write",
@@ -13,10 +13,10 @@
     "crm": {
       "cards": [
         {
-          "file": "extensions/stage-tracker.json"
+          "file": "extensions/stage-tracker-example-card.json"
         },
         {
-          "file": "extensions/association-table.json"
+          "file": "extensions/association-table-example-card.json"
         }
       ]
     }
