+++ crm-data-components/src/app/extensions/package.json
