+++ crm-data-components/src/app/extensions/stage-tracker-example-card.json
@@ -1,8 +1,8 @@
 {
   "type": "crm-card",
-  "uid": "stage_tracker_extension",
+  "uid": "stage_tracker_example",
   "data": {
-    "title": "Stage Tracker Card",
+    "title": "CRM data: Stage Tracker example card",
     "location": "crm.record.tab",
     "module": {
       "file": "StageTrackerExtension.tsx"
